mise en place controllers

+++ src/main/java/fr/cesi/ril18/CineGoBack/repositories/SceanceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import fr.cesi.ril18.CineGoBack.entities.Sceances;
 
-public interface SceanceRepository extends CrudRepository<Sceances, Long>{
+public interface SceanceRepository extends CrudRepository<Sceances, Integer>{
 
 	
 	List<Sceances> findAll();
